+++ packages/swc-plugin/tests/index.test.ts
@@ -20,6 +20,7 @@ const FILES_TO_TESTS: Record<string, Opts> = {
   extractSourceLocation: {
     extractSourceLocation: true,
   },
+  descriptionsAsObjects: {},
   formatMessageCall: {},
   FormattedMessage: {},
   inline: {},